112 move test classes to folders
