+++ test/Kentico.Kontent.ModelGenerator.Tests/Common/PropertyTests.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using Kentico.Kontent.ModelGenerator.Core;
 using Kentico.Kontent.ModelGenerator.Core.Common;
 using Xunit;
 
-namespace Kentico.Kontent.ModelGenerator.Tests
+namespace Kentico.Kontent.ModelGenerator.Tests.Common
 {
     public class PropertyTests
     {
