+++ test/Kentico.Kontent.ModelGenerator.Tests/Helpers/ElementIdHelperTests.cs
@@ -4,12 +4,11 @@ using Kentico.Kontent.Delivery.Abstractions;
 using Kentico.Kontent.Management.Models.Types;
 using Kentico.Kontent.Management.Models.Types.Elements;
 using Kentico.Kontent.Management.Models.TypeSnippets;
-using Kentico.Kontent.ModelGenerator.Core;
 using Kentico.Kontent.ModelGenerator.Core.Helpers;
 using Newtonsoft.Json.Linq;
 using Xunit;
 
-namespace Kentico.Kontent.ModelGenerator.Tests
+namespace Kentico.Kontent.ModelGenerator.Tests.Helpers
 {
     public class ElementIdHelperTests
     {
